backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
motor==3.3.2
httpx==0.25.2
solana==0.30.2
solders==0.18.1
base58==2.1.1
requests==2.31.0
Pillow==10.1.0
aiofiles==23.2.1